Fix condensebook pruning threshold truncation bug
Integer truncation of max_visits * min_ratio allowed moves well below
the intended 10% cutoff to survive (e.g. 2.9% probability moves).
Use double threshold to avoid truncation.

+++ condensebook.cpp
@@ -117,7 +117,7 @@ std::vector<CondensedMove> condense_entry(const BookEntry& entry,
 
   if (max_visits < min_visits) return {};
 
-  int threshold = static_cast<int>(max_visits * min_ratio);
+  double threshold = max_visits * min_ratio;
 
   std::vector<CondensedMove> result;
   int filtered_total = 0;
